Add files via upload
ASCII-ART-COLOR

This Golang code turns ascii characters into images, and applies colour to those images. The colours have been achieved by using the RGB notation.

The user is able to choose between colouring a single letter, a set of contiguous or non-contiguous letters, or the entire word(s). The output is determined by manipulating the 'flag' terminal command:  --color=<color> _<index>, in which --color is the flag, <color> is the color desired by the user, and  _<index> determines the position of letters to be coloured. If no index is provided, the entire word will be coloured.

For example: 

 --color=blue_1:4 

will apply blue colour to the second, third and fourth characters. Note that indexes for word character are zero-based.